Remove old export macro
The previous macro was impossible to maintain because it fused exportation logic and pre exportation logic. The new macro is better as it clearly divides both tasks, so upgrading to an up to date sophia version should be less painfull. Also, this new macro let the user redefine some functions more easily. The only drawback is that we can't produce another kind of exported dataset that the one we are calling the match_quad function on, but I do think it was a bad design decision to solve the problem it tried to address (filling a new dataset as fast as possible)